+++ CHANGELOG.md
@@ -20,12 +20,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `v0.0.3` CopyClipboardInfoCard
   - Information span overflow & ellipsis
 
-## 0.19.2 - 2024-07-22
 ### Added
+
+- `v0.0.1` Checkbox component
+- `v0.0.3` ComboBox & `v0.0.2` ComboBoxButton
+  - add optional `showIcon` property (set default to `true`)
+  - add button shadow on enabled options
+
+## 0.19.2 - 2024-07-22
+
+### Added
+
 - New icons following `front-tokens@0.1.19` update
 
 ## 0.19.1 - 2024-07-18
+
 ### Fixed
+
 - `v0.0.2` CopyClipboardInfoCard
   - Display underline in href label even through ellipsis
   - Change href label color on hover
@@ -35,7 +46,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Use `@storybook/builder-vite` in config
 
 ## 0.19.0 - 2024-07-17
+
 ### Added
+
 - `v0.0.1` CopyClipboardInfoCard
   - Displays an info and / or a link in an easily readable card
   - Allows easy copying to clipboard
@@ -44,6 +57,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Renders several levels of data and info in a easily digestible manner
 
 ### Changed
+
 - `v0.0.2` Switch component
   - styles: disabled colors to match Figma updates on the component
   - add `withHelper`, `activeHelperText` and `inactiveHelperText` to handle the displaying of the text
@@ -51,34 +65,46 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Allow cursor customization, defaults to `help`
 
 ## 0.18.0 - 2024-07-16
+
 ### Added
+
 - `v0.0.1` Switch component
 - `v0.0.4` Text component
   - an optional uppercase property to capitalize its text contents
 
 ## 0.17.1 - 2024-07-09
+
 ### Fixed
+
 - Input / Field controlled / uncontrolled states and value propagation
 
 ## 0.17.0 - 2024-07-08
+
 ### Added
+
 - `v0.0.1` TextAreaInput component
 
 ### Fixed
+
 - (Un)controlled input detection & warnings
 
 ## 0.16.1 - 2024-07-08
+
 ### Changed
+
 - `v0.0.3` ComboBox
   - Support for setting & updating selection through `value` prop
 
 ### Fixed
+
 - `v0.0.4` Input components
   - replace placholder `placeholder` default prop with empty string
 - TYPES: Do not export inner flex layout to avoid ts error when used as a dependency
 
 ## 0.16.0 - 2024-07-02
+
 ### Added
+
 - `v0.0.1` Divider component
   - A horizontal line with an optional label
 - `v0.0.1` Form component
@@ -87,6 +113,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Data validation using zod schemas, error propagation
 
 ### Changed
+
 - Input styles
   - Added support for displaying inside a table cell
 - `v0.0.2` GridItem layout
@@ -97,94 +124,127 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Added support for object / array props with proper formatting
 
 ### Fixed
+
 - `v0.0.4` Field component
   - Do not render assistive / error text if no content available
   - This prevents rendering an additional gap at the bottom of the field between its input and an effectively empty div.
 
 ## 0.15.0 - 2024-07-02
+
 ### Added
+
 - `v0.0.1` Avatar component
 
 ### Changed
+
 - `v0.0.4` EmailInput component
   - added "autocomplete" attribute
 
 ## 0.14.1 - 2024-06-27
+
 ### Fixed
+
 - `v0.0.2` ComboBox component
   - include "combobox" ARIA role
 - `v0.0.2` ComboBoxButton component
   - set type to "button"
 
 ## 0.14.0 - 2024-06-27
+
 ### Added
+
 - `v0.0.1` ComboBox component
 - `v0.0.1` ComboBoxButton component
 
 ## 0.13.0 - 2024-06-26
+
 ### Added
+
 - `v0.0.1` Modal component
 
 ## 0.12.0 - 2024-06-24
+
 ### Added
+
 - `v0.0.1` InfoBox component
 
 ## 0.11.2 - 2024-06-20
+
 ### Changed
+
 - Upgraded storybook to latest version
 
 ### Fixed
+
 - `v0.0.3` EmailInput component
   - Use standard validation regex
 - `v0.0.4` PasswordInput component
   - Include `autoComplete="new-password"` attribute
 
 ## 0.11.1 - 2024-06-19
+
 ### Fixed
+
 - `v0.0.4` TextInput component
   - TestIdProps
 
 ## 0.11.0 - 2024-06-19
+
 ### Added
+
 - `v0.0.2` Loading component
-  
+
   - animation type `ProgressBar`
   - `className` & `TestId` props
+
 - `v0.0.1` SmallLoader component
 
 ### Fixed
+
 - `v0.0.4` NumberInput component
-  
+
   - prevent controls from changing value when input is `disabled`
+
 - `v0.0.4` StateIndicator component
   - styles: set overflow to hidden to disable scrollbar inside the component in Firefox
 
 ## 0.10.8 - 2024-06-17
+
 ### Added
+
 - `v0.0.6` Button components
   - `fullWidth` prop: fills the parent container if true
 
 ## 0.10.7 - 2024-06-13
+
 ### Fixed
+
 - `v0.0.5` Button component
   - Avoid text break by setting inner span width to `max-content`
 
 ## 0.10.6 - 2024-06-13
+
 ### Changed
+
 - `v0.0.4` Button component
   - Add testid prop support
   - Add style override usage (`styled(Button)`) support
 
 ### Fixed
+
 - reduced all input components `min-width` to 6rem
 
 ## 0.10.5 - 2024-06-12
+
 ### Changed
+
 - `v0.0.6` Badge component
   - increased `border-radius` for rounded corners
 
 ## 0.10.4 - 2024-06-11
+
 ### Added
+
 - `onBlur` & `inputRef` to all input & field components
 - `useInputRef` custom hook
   - used to assign an element to both parent and inner `inputRef`s
@@ -193,28 +253,37 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - normalizes an nullable value and returns `undefined` if `null`
 
 ## 0.10.3 - 2024-06-06
+
 ### Changed
+
 - `v.0.0.5` Badge component
   - Require at least one of `label` or `icon`
 
 ## 0.10.2 - 2024-06-06
+
 ### Fixed
+
 - `v.0.0.4` Chip component
   - assign onDelete to delete button onMouseDown as well to stop event propagation
 
 ## 0.10.1 - 2024-06-06
+
 ### Fixed
+
 - `v.0.0.3` Chip component
   - stop onDelete event propagation
 
 ## 0.10.0 - 2024-06-06
+
 ### Added
+
 - `v0.0.1` CurrencyInput component
 - New `name` & `onChangeNative` props for all Inputs & Field components
 - `useTestId` & `TestIdProps`
   - Ability to override / replace default test id
 
 ### Changed
+
 - Propagate native input event from custom inputs regardless of onChange condition
 - `v0.0.2` Field component
   - Display generic error messages based on underlying input validity
@@ -225,26 +294,34 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Add support for custom test id
 
 ### Fixed
+
 - `v.0.0.2` Chip component
   - make icon property optional
 
 ## 0.9.0 - 2024-06-04
+
 ### Added
+
 - `v0.0.1` Chip component
 
 ## 0.8.1 - 2024-06-03
+
 ### Fixed
+
 - `v0.0.3` State Indicator
   - Use correct text size (from S to M)
 
 ## 0.8.0 - 2024-06-03
+
 ### Added
+
 - `v0.0.1` Field component
 - New `className` prop for compatibility with `styled([Component])` usage:
   - `Text`, `Heading` typo components
   - `Flex`, `Grid` & `GridItem` layouts
 
 ### Changed
+
 - Component init script
   - Added Boilerplace file contents
 - Generic Input component typings
@@ -252,41 +329,56 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Added boilerplate content
 
 ### Fixed
+
 - Various minor fixes in multiple components
   - Unit tests
   - Styles
   - Test ids
 
 ## 0.7.0 - 2024-05-30
+
 ### Added
+
 - `v0.0.1` Tooltip component
 
 ## 0.6.4 - 2024-05-29
+
 ### Fixed
+
 - Wait for overrides to have been fetched before applying favicon
 
 ## 0.6.3 - 2024-05-28
+
 ### Fixed
+
 - Compare favicon retry count to retry limit, allowing favicon to actually be applied
 
 ## 0.6.2 - 2024-05-28
+
 ### Fixed
+
 - Actually check if favicon has been applied
 
 ## 0.6.1 - 2024-05-28
+
 ### Added
+
 - Set favicon from theme provider according to overrides
 
 ### Changed
+
 - Theme provider optimizations
 - Bumped `@ubloimmo/front-tokens` to `0.0.17`
 - Rewrite `GlobalStyle` in order to decrease excessive renders and increase perf
 
 ### Fixed
+
 - Effect color mapping thanks to updated design tokens
 
 ## 0.6.0 - 2024-05-28
+
 ### Added
+
 - `v0.0.1` GridItem layout
 - Docs
   - `Typography` page
@@ -295,6 +387,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - Link to related `Properties` page
 
 ### Changed
+
 - Story sorting in storybook
 - Docs
   - Index pages
@@ -302,25 +395,33 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Page Header Styles
 
 ### Fixed
+
 - Shadow / effect parsing following `@ubloimmo/front-tokens` regen
 - Various minor fixes
 
 ## 0.5.1 - 2024-05-23
+
 ### Fixed
+
 - `v0.0.3` PhoneInput component
   - width
   - z-clashing
   - responsive height
 
 ## 0.5.0 - 2024-05-22
+
 ### Added
+
 - `v.0.0.1` Hypertext component
 
 ## 0.4.4 - 2024-05-21
+
 ### Added
+
 - Phone input default placeholder
 
 ### Fixed
+
 - Phone input foreign styles
 
 ## 0.4.3 - 2024-05-21
@@ -330,20 +431,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## 0.4.1 - 2024-05-21
 
 ## 0.4.0 - 2024-05-20
+
 ### Added
+
 - `v.0.0.1` PhoneInput component
 - Full coverage test for NumberInput component
 
 ## 0.3.2 - 2024-05-14
+
 ### Fixed
+
 - Theme override validation condition
 
 ## 0.3.1 - 2024-05-14
+
 ### Fixed
+
 - Theme override URL template
 
 ## 0.3.0 - 2024-05-14
+
 ### Added
+
 - `v0.0.1` Action Icon component
   - Logic
   - Styles
@@ -352,33 +461,42 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Extract` type: a safeguard when dealing with union subsets
 
 ### Fixed
+
 - `v0.0.2` State Indicator component
   - Set minimum height to 40px
   - Reduce markup complexity (remove nested div)
 
 ## 0.2.0 - 2024-05-13
+
 ### Added
+
 - `v.0.0.1` State indicator component
 - `EnumExtension` generic type
 
 ### Changed
+
 - Typings
   - Replace `StaticIconColor` with `ColorKeyOrWhite`
 
 ### Fixed
+
 - Text & Heading stories text content
 
 ## 0.1.2 - 2024-04-25
 
 ## 0.1.1 - 2024-04-17
+
 ### Fixed
+
 - `v0.0.3` Static Icon component
   - Sizing
 - `v0.0.2` Action component
   - Responsive behaviour
 
 ## 0.1.0 - 2024-04-16
+
 ### Added
+
 - License & copyright
 - `v0.0.1` Action component
   - [Usage](https://ublo.ublo.org/design-system?path=/docs/components-action-usage--docs)
@@ -390,6 +508,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Card elevation shadow
 
 ### Changed
+
 - `v0.0.2` Button component
   - Secondary styles (white, primary, error)
 - `v0.0.3` Badge component
@@ -398,10 +517,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - White variant
 
 ### Removed
+
 - Icons:
   - Delete (custom icon)
 
 ### Fixed
+
 - `v0.0.2` InputLabel
   - Text size after font style overhaul
 - `v0.0.2` InputAssistiveText
@@ -410,7 +531,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## 0.1.0-beta.0 - 2024-04-11
 
 ## 0.0.2-beta.0 - 2024-04-11
+
 ### Added
+
 - Global Theme
 - Theme context, provider & consumer hook.
 - Global style injector (based on `styled-components`'s `GlobalStyle` component).